requirements.txt
flask
Click==7.0
itsdangerous==1.1.0
lxml==4.4.2
python-docx==0.8.10
Werkzeug==0.16.0